refactor(cli): restructure as Param/Arg with Ts.Err integration
Major refactor to clarify conceptual separation and modernize error handling.

**Conceptual Separation:**
- Cli.Param: Parameter specifications (what CLI accepts) - was FlagName
- Cli.Arg: Runtime arguments (what user types) - was ArgToken

**Structure Changes:**
- Consolidated into monolithic files (no separate .types.ts)
- All types inline in TypeScript namespaces (declaration merging)
- Compact table-style tests using Test.on()

**Error Handling:**
- Migrated to Ts.Err.StaticError for structured type-level errors
- Removed StaticErrorAny (use StaticErrorLike for constraints)
- Better DX with metadata and IDE hover experience

**Files:**
- cli/param/param.ts - Parameter specs (631 lines)
- cli/param/param.test.ts - Compact tests (21 tests)
- cli/arg/arg.ts - Runtime args (500 lines)
- cli/arg/arg.test.ts - Compact tests (21 tests)

All 116 tests passing ✅
