--- src/utils/cli/arg-token/$.ts
@@ -1,1 +1,0 @@
-export { ArgToken } from './arg-token.js'
