--- src/utils/cli/flag-name/$.ts
@@ -1,1 +1,0 @@
-export { FlagName } from './flag-name.js'
